Update patch set 1
Patch Set 1:

Was testing a few things to see if battery usage was improved. I think I originally grabbed the change from Mako, but I notice it's not enabled in the config file for the kk_2.7 branch. Kinda forgot to complete the testing with and without it enabled, so thanks for reminding me. :)

Patch-set: 1