Update patch set 1
Patch Set 1:

Hmm, seems to be the same in Google camera with and without this patch. When taking HDR pictures, the app randomly crashes. I'm not even sure HDR is working properly, as I can't see any differences in pictures taken with and without it. There's another problem with Google camera app when using LED flash - sometimes the LED get stuck on forever and it doesn't capture the image. I have to switch to camera mode and back in order for the LED to turn off.

Patch-set: 1